simple Makefile for installing puppet modules
install available puppet modules into /usr/share/puppet/modules
i use this so i can then puppet apply -e "include <classname>"
for now. i plan on replacing this with a local node setup when
it contains more components that work, but i am editing this
commit message with a vim that was installed using this.