started some predefined config for mixxx
i want to keep as much as possible in puppet so im starting to add
core config templates that contain stripped down config file. this
means that you will have to look up details on config sytax that
are not relevant to us in info/man pages or on a setup without
stripped setup files. personally one could argue in both directions
on this decision, for now the files are stripped so our that there
is more than enough room for me to add documentation that is relevant
to the problems being solved here.