install and configure jack/alsa
this is still mostly just a standrd config of the whole jack/alsa
shebang wihtout much optimization. it is setup so the whole thing
can be started as a service by init since i believe that to be the
most stable way for running jack on a single purpose machine like
this.