install kernel sources and tools
This is where i install kernel sources and install genkernel. It
contains some configuration for the two, but does now build a kernel
by itself as of now. I plan on adding some more automation to this
part of the configuration process, but i believe that we will probably
end up using a self-but-pre-build binary kernel of some sort for this
project.